Split up azure-account.ts into multiple files
